split out database types and constants from init code
