+++ src/components/Draft.tsx
@@ -12,8 +12,8 @@ import RadioButtonUncheckedIcon from "@mui/icons-material/RadioButtonUnchecked";
 import { CheckCircleTwoTone as Check } from "@mui/icons-material";
 import { styled } from "@mui/material/styles";
 import { useUpdateAnimation } from "../hooks/useUpdateAnimation";
-import { GBGameStateDoc } from "../models/gbdb";
 import { Model, Guild } from "../components/DataTypes";
+import { GBGameStateDoc } from "../models/gbdbTypes";
 import { reSort } from "../utils/reSort";
 import { useRxData } from "../hooks/useRxQuery";
 import { map, Observable } from "rxjs";
